requirements.txt
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
scikit-surprise==1.1.1
flask==2.3.2
